+++ packages/snap-client/src/Client/apis/Recommend.ts
@@ -157,6 +157,9 @@ export class RecommendAPI extends API {
 				const response = await this.postRecommendations(batch.request as RecommendPostRequestModel);
 
 				batch.entries?.forEach((entry, index) => {
+					response[index]?.results?.forEach((result, idx) => {
+						result.position = idx + 1;
+					});
 					entry.deferred.resolve([response[index]]);
 				});
 			} catch (err) {