+++ packages/snap-preact-demo/tests/cypress/e2e/tracking/track.cy.js
@@ -55,134 +55,117 @@ describe('Tracking Beacon 2.0', () => {
 
 	it('tracked search render, impression, clickthrough', () => {
 		cy.visit('https://localhost:2222');
-		let initialContext;
-		cy.waitForBundle().then(() => {
-			cy.snapController().then((controller) => {
-				const results = controller.store.results.filter((result) => result.type === 'product');
-				expect(results.length).to.be.greaterThan(0);
+		new Cypress.Promise.all([cy.wait(`@beacon2/search/render`), cy.wait(`@beacon2/search/impression`)])
+			.then(([render, impression]) => {
+				expect(render.response.body).to.have.property('success').to.equal(true);
+				const { context: context1, data: data1 } = JSON.parse(render.request.body);
+				expect(context1).to.be.an('object');
+				expect(data1).to.have.property('results').to.be.an('array');
+				cy.get('.ss__result').should('have.length', data1.results.length); // all results are rendered
+				expect(data1.results[0]).to.have.property('uid').to.be.a('string');
+				expect(data1).to.have.property('merchandising').to.be.an('object');
+				expect(data1).to.have.property('pagination').to.be.an('object');
+				expect(data1).to.not.have.property('sort');
+				expect(data1).to.have.property('q').to.be.a('string');
 
-				cy.wait(`@beacon2/search/render`)
-					.then(({ request, response }) => {
-						expect(response.body).to.have.property('success').to.equal(true);
+				expect(impression.response.body).to.have.property('success').to.equal(true);
+				const { data: data2, context: context2 } = JSON.parse(impression.request.body);
+				expect(data2).to.have.property('results').to.be.an('array').length(4); // first 4 results are visible on the page
+				// assert context values are the same as the initial values
+				expect(context2).to.be.an('object');
+				expect(context2.initiator).to.equal(context1.initiator);
+				expect(context2.pageLoadId).to.equal(context1.pageLoadId);
+				expect(context2.pageUrl).to.equal(context1.pageUrl);
+				expect(context2.sessionId).to.equal(context1.sessionId);
+				expect(context2.shopperId).to.equal(context1.shopperId);
+				expect(context2.timestamp).not.to.equal(context1.timestamp); // timestamp should be different
+				expect(context2.userAgent).to.equal(context1.userAgent);
+				expect(context2.userId).to.equal(context1.userId);
+				expect(context2.pageLoadId).to.equal(context1.pageLoadId);
+				return { context1 }; // return the context values to be used in the next test
+			})
+			.then(({ context1 }) => {
+				// reload page to generate new context
+				cy.visit('https://localhost:2222?differentPageUrl=1');
+				cy.waitForBundle().then(() => {
+					cy.snapController().then((controller) => {
+						const firstResult = controller.store.results.find((result) => result.type === 'product');
+						const pagination = controller.store.pagination;
+						const merchandising = controller.store.merchandising;
+						expect(firstResult).to.exist;
+						expect(pagination).to.exist;
+						expect(merchandising).to.exist;
 
-						const { context, data } = JSON.parse(request.body);
-						expect(context).to.be.an('object');
+						cy.wait(1000);
+						// Click on first product
+						cy.get(`.ss__result a[href='${firstResult.mappings.core.url}']`).first().click({ force: true });
 
-						// save context values for later assertions
-						initialContext = { ...context };
+						cy.wait(`@beacon2/search/clickthrough`).then((clickthrough) => {
+							expect(clickthrough.response.body).to.have.property('success').to.equal(true);
 
-						expect(data).to.have.property('results').to.be.an('array').length(results.length); // all results are rendered
-						expect(data.results[0]).to.have.property('uid').to.be.a('string');
-						expect(data).to.have.property('merchandising').to.be.an('object');
-						expect(data).to.have.property('pagination').to.be.an('object');
-						expect(data).to.not.have.property('sort');
-						expect(data).to.have.property('q').to.be.a('string');
-					})
-					.then(() => {
-						cy.wait(`@beacon2/search/impression`).then(({ request, response }) => {
-							expect(response.body).to.have.property('success').to.equal(true);
+							const { context: context3, data: data3 } = JSON.parse(clickthrough.request.body);
 
-							const { context, data } = JSON.parse(request.body);
-							expect(data).to.have.property('results').to.be.an('array').length(4); // first 4 results are visible on the page
+							// assert clickthrough data is correct product that was clicked
+							expect(data3).to.have.property('results').to.be.an('array').length(1);
+							expect(data3.results[0]).to.have.property('uid').to.be.a('string').and.to.equal(firstResult.mappings.core.uid);
+							expect(data3.results[0]).to.have.property('sku').to.be.a('string').and.to.equal(firstResult.mappings.core.sku);
 
-							// assert context values are the same as the initial values
-							expect(context).to.be.an('object');
-							expect(context.initiator).to.equal(initialContext.initiator);
-							// expect(context.pageLoadId).to.equal(initialContext.pageLoadId);
-							expect(context.pageUrl).to.equal(initialContext.pageUrl);
-							expect(context.sessionId).to.equal(initialContext.sessionId);
-							expect(context.shopperId).to.equal(initialContext.shopperId);
-							expect(context.timestamp).not.to.equal(initialContext.timestamp); // timestamp should be different
-							expect(context.userAgent).to.equal(initialContext.userAgent);
-							expect(context.userId).to.equal(initialContext.userId);
-							expect(context.pageLoadId).to.equal(initialContext.pageLoadId);
-						});
-					})
-					.then(() => {
-						// reload page to generate new context
-						cy.visit('https://localhost:2222?differentPageUrl=1');
+							// assert data doesn't have other properties
+							expect(JSON.stringify(data3)).to.equal(
+								JSON.stringify({
+									q: '',
+									matchType: 'primary',
+									pagination: {
+										totalResults: pagination.totalResults,
+										page: pagination.page,
+										resultsPerPage: pagination.pageSize,
+									},
+									merchandising: {
+										redirect: merchandising.redirect,
+										personalized: merchandising.personalized,
+									},
+									results: [
+										{
+											position: 1,
+											uid: firstResult.mappings.core.uid,
+											sku: firstResult.mappings.core.sku,
+										},
+									],
+								})
+							);
 
-						cy.waitForBundle().then(() => {
-							cy.snapController().then((controller) => {
-								const firstResult = controller.store.results.find((result) => result.type === 'product');
-								const pagination = controller.store.pagination;
-								const merchandising = controller.store.merchandising;
-								expect(firstResult).to.exist;
-								expect(pagination).to.exist;
-								expect(merchandising).to.exist;
-
-								// Click on first product
-								cy.get(`.ss__result a[href='${firstResult.mappings.core.url}']`).first().click({ force: true });
-
-								cy.wait(`@beacon2/search/clickthrough`).then(({ request, response }) => {
-									expect(response.body).to.have.property('success').to.equal(true);
-
-									const { context, data } = JSON.parse(request.body);
-
-									// assert clickthrough data is correct product that was clicked
-									expect(data).to.have.property('results').to.be.an('array').length(1);
-									expect(data.results[0]).to.have.property('uid').to.be.a('string').and.to.equal(firstResult.mappings.core.uid);
-									expect(data.results[0]).to.have.property('sku').to.be.a('string').and.to.equal(firstResult.mappings.core.sku);
-
-									// assert data doesn't have other properties
-									expect(JSON.stringify(data)).to.equal(
-										JSON.stringify({
-											q: '',
-											pagination: {
-												totalResults: pagination.totalResults,
-												page: pagination.page,
-												resultsPerPage: pagination.pageSize,
-											},
-											merchandising: {
-												redirect: merchandising.redirect,
-												personalized: merchandising.personalized,
-											},
-											results: [
-												{
-													uid: firstResult.mappings.core.uid,
-													sku: firstResult.mappings.core.sku,
-												},
-											],
-										})
-									);
-
-									// assert context values are different after page reload
-									expect(context).to.be.an('object');
-									expect(context.initiator).to.equal(initialContext.initiator);
-									expect(context.pageLoadId).to.not.equal(initialContext.pageLoadId); // pageLoadId should be different
-									expect(context.pageUrl).to.not.equal(initialContext.pageUrl); // pageUrl should be different
-									expect(context.sessionId).to.equal(initialContext.sessionId);
-									expect(context.shopperId).to.equal(initialContext.shopperId);
-									expect(context.timestamp).not.to.equal(initialContext.timestamp); // timestamp should be different
-									expect(context.userAgent).to.equal(initialContext.userAgent);
-									expect(context.userId).to.equal(initialContext.userId);
-								});
-							});
+							// assert context values are different after page reload
+							expect(context3).to.be.an('object');
+							expect(context3.initiator).to.equal(context1.initiator);
+							expect(context3.pageLoadId).to.not.equal(context1.pageLoadId); // pageLoadId should be different
+							expect(context3.pageUrl).to.not.equal(context1.pageUrl); // pageUrl should be different
+							expect(context3.sessionId).to.equal(context1.sessionId);
+							expect(context3.shopperId).to.equal(context1.shopperId);
+							expect(context3.timestamp).not.to.equal(context1.timestamp); // timestamp should be different
+							expect(context3.userAgent).to.equal(context1.userAgent);
+							expect(context3.userId).to.equal(context1.userId);
 						});
 					});
+				});
 			});
-		});
 	});
 
 	it('tracked category render, impression, clickthrough', () => {
 		cy.visit('https://localhost:2222/category.html');
 
-		cy.wait(`@beacon2/category/render`).then(({ request, response }) => {
-			expect(response.body).to.have.property('success').to.equal(true);
-
-			const { context, data } = JSON.parse(request.body);
+		new Cypress.Promise.all([cy.wait(`@beacon2/category/render`), cy.wait(`@beacon2/category/impression`)]).then(([render, impression]) => {
+			expect(render.response.body).to.have.property('success').to.equal(true);
+			const { context, data } = JSON.parse(render.request.body);
 			expect(context).to.be.an('object');
 			expect(data).to.have.property('results').to.be.an('array').length.greaterThan(0);
 			expect(data).not.to.have.property('q');
-
 			expect(data).to.have.property('bgfilter').to.be.an('array').length.greaterThan(0);
-		});
 
-		cy.wait(`@beacon2/category/impression`, { timeout: 10000 }).its('response.statusCode').should('eq', 200);
+			expect(impression.response.statusCode).to.equal(200);
+		});
 
 		// Click on first product
 		cy.get('.ss__result a').first().click({ force: true });
-
 		cy.wait(`@beacon2/category/clickthrough`).its('response.statusCode').should('eq', 200);
 	});
 
@@ -190,23 +173,19 @@ describe('Tracking Beacon 2.0', () => {
 		cy.visit('https://localhost:2222');
 		cy.get('input[name="q"]').type('s');
 
-		cy.wait(`@beacon2/autocomplete/render`).then(({ request, response }) => {
-			expect(response.body).to.have.property('success').to.equal(true);
-
-			const { context, data } = JSON.parse(request.body);
+		new Cypress.Promise.all([cy.wait(`@beacon2/autocomplete/render`), cy.wait(`@beacon2/autocomplete/impression`)]).then(([render, impression]) => {
+			expect(render.response.body).to.have.property('success').to.equal(true);
+			const { context, data } = JSON.parse(render.request.body);
 			expect(context).to.be.an('object');
 			expect(data).to.have.property('results').to.be.an('array').length.greaterThan(0);
 			expect(data.results[0]).to.have.property('uid').to.be.a('string');
 			expect(data).to.have.property('merchandising').to.be.an('object');
 			expect(data).to.have.property('pagination').to.be.an('object');
 			expect(data).to.not.have.property('sort');
-		});
 
-		cy.wait(`@beacon2/autocomplete/impression`, { timeout: 10000 }).then(({ request, response }) => {
-			expect(response.body).to.have.property('success').to.equal(true);
-
-			const { data } = JSON.parse(request.body);
-			expect(data).to.have.property('results').to.be.an('array').length.greaterThan(0);
+			expect(impression.response.body).to.have.property('success').to.equal(true);
+			const { data: data2 } = JSON.parse(impression.request.body);
+			expect(data2).to.have.property('results').to.be.an('array').length.greaterThan(0);
 		});
 
 		// Click on first product