+++ packages/snap-store-mobx/src/Autocomplete/Stores/AutocompleteQueryStore.ts
@@ -1,12 +1,17 @@
 import { observable, makeObservable } from 'mobx';
 import type { UrlManager } from '@searchspring/snap-url-manager';
-import type { SearchResponseModelSearch, AutocompleteResponseModelAllOfAutocomplete } from '@searchspring/snapi-types';
+import type {
+	SearchResponseModelSearch,
+	AutocompleteResponseModelAllOfAutocomplete,
+	SearchResponseModelSearchMatchTypeEnum,
+} from '@searchspring/snapi-types';
 import type { AutocompleteStoreConfig, StoreServices } from '../../types';
 
 export class AutocompleteQueryStore {
 	public query?: Query;
 	public originalQuery?: Query;
 	public correctedQuery?: Query;
+	public matchType: SearchResponseModelSearchMatchTypeEnum;
 
 	constructor(
 		services: StoreServices,
@@ -31,6 +36,9 @@ export class AutocompleteQueryStore {
 			}
 		}
 
+		this.matchType = search.matchType as SearchResponseModelSearchMatchTypeEnum;
+		observables.matchType = observable;
+
 		makeObservable(this, observables);
 	}
 }
@@ -39,6 +47,7 @@ type Observables = {
 	originalQuery?: typeof observable;
 	query?: typeof observable;
 	correctedQuery?: typeof observable;
+	matchType?: typeof observable;
 };
 
 class Query {